requirements.txt
django
Pillow
requests
